Add utils and cleanup a bit
